requirements.txt
requests~=2.26.0
redis~=3.5.3
Flask~=2.0.1
ping3~=3.0.2
python-dotenv~=0.19.0
websocket-client~=1.2.1
click~=7.1.2
rq~=1.10.0
pip~=21.2.4
rsa~=4.7.2
pyasn1~=0.4.8
vine~=5.0.0
amqp~=5.0.6
pytz~=2021.1
PyYAML~=5.4.1
MarkupSafe~=2.0.1
Werkzeug~=2.0.1
Jinja2~=3.0.1
itsdangerous~=2.0.1
kombu~=5.1.0
billiard~=3.6.4.0
cachetools~=4.2.2
certifi~=2021.5.30
urllib3~=1.26.7
six~=1.16.0
wcwidth~=0.2.5
croniter~=1.0.15
python-dateutil~=2.8.2
gunicorn~=20.1.0
oauthlib~=3.1.1
idna~=3.2
kubernetes~=18.20.0
setuptools~=41.2.0